:art: LocationHandler 제거 및 리팩토링

+++ feature/footprint/src/main/java/com/umc/footprint/component/card/DiaryCardHorizontalPageArrow.kt
@@ -36,12 +36,12 @@ fun DiaryCardHorizontalPageArrow(
 
     val outerColor by animateColorAsState(
         targetValue = prop.outerColor,
-        animationSpec = tween(durationMillis = prop.colorAnimationDuration.toInt()),
+        animationSpec = tween(durationMillis = prop.colorAnimationDuration),
     )
 
     val innerColor by animateColorAsState(
         targetValue = prop.innerColor,
-        animationSpec = tween(durationMillis = prop.colorAnimationDuration.toInt()),
+        animationSpec = tween(durationMillis = prop.colorAnimationDuration),
     )
 
     Canvas(
